+++ jekyll-theme-cicse/CHANGELOG.md
@@ -10,10 +10,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Theme changelog
+- Constructor corporate colors for Tailwind:
+  ```javascript
+  // Main colors
+  'mobility-blue': '#008ce3',
+  'navy-blue': '#00204d',
+  'diversity-red': '#db4f3d',
+  // Accent colors
+  'confident-blue': '#004381',
+  'shiny-yellow': '#ffc622',
+  'healthy-green': '#00a987',
+  'graphene-black': '#0a0a0a'
+  ```
 
 ### Changed
 
 - Use a different style for collection pages which have `is_index`
+- Use `navy-blue` as the main color
 
 ### Removed